+++ lib/commands/analyze-report.ts
@@ -39,6 +39,7 @@ export async function handler(argv: yargs.Arguments): Promise<void> {
       newmanReportFilePath: newmanReportPath,
       reportOutputFilePath: reportOutputFilePath,
       validationLevel: argv.level,
+      generateExample: true,
     };
     const analyzer = inversifyGetInstance(NewmanReportAnalyzer, opts);
     await analyzer.analyze();