+++ test/apiScenario/apiScenarioLoaderTest.ts
@@ -20,6 +20,25 @@ describe("ApiScenarioLoader", () => {
     expect(testDef).toMatchSnapshot();
   });
 
+  it("load valid example based scenario - storage", async () => {
+    const fileRoot = "test/apiScenario/fixtures/specification/storage/resource-manager/";
+
+    const loader = ApiScenarioLoader.create({
+      useJsonParser: false,
+      eraseXmsExamples: false,
+      checkUnderFileRoot: false,
+      fileRoot,
+      swaggerFilePaths: ["Microsoft.Storage/stable/2021-08-01/storage.json"],
+      includeOperation: false,
+    });
+
+    const testDef = await loader.load(
+      "Microsoft.Storage/stable/2021-08-01/scenarios/storageBasicExample.yaml"
+    );
+
+    expect(testDef).toMatchSnapshot();
+  });
+
   it.skip("load valid cross-RP scenario - compute", async () => {
     const fileRoot = "test/apiScenario/fixtures/specification/compute/resource-manager/";
 
@@ -76,7 +95,7 @@ describe("ApiScenarioLoader", () => {
 
   it("load valid scenario from uri", async () => {
     const readmeMd: string =
-      "https://github.com/Azure/azure-rest-api-specs/blob/apiscenario/specification/signalr/resource-manager/readme.md";
+      "https://github.com/Azure/azure-rest-api-specs/blob/23117e9/specification/signalr/resource-manager/readme.md";
     const fileRoot = pathDirName(readmeMd);
 
     const loader = ApiScenarioLoader.create({