+++ ChangeLog.md
@@ -12,6 +12,7 @@
   - Support running API Scenario against localhost endpoint
   - Refactor Postman based runner client
   - Support API Scenario generation from testProxy recording
+  - Support local devMode where AAD auth, ARM call and LRO polling will be disabled
 
 ## 05/27/2022 2.12.2